Update .gitignore to track specific address files and municipality_code.csv
